be/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain-core
langgraph
qdrant-client
streamlit
requests
langchain-gigachat
langchain-community
